freelist stopped compiling for some weird unknown reason... revert when I get a time to test if the correct code has started working again BindingTable and BindingTableMemoryMap done upto inserting into a pipeline (todo actual writing the binding data) Fix for Image in placeInStorage Refactor names in Resource manager, dropped Resource in calls except when actually getting a resource. Added a one shot resourceFromName function Pipeline wip - ROPBlender and Viewport added

+++ libs/vulkan/commandqueue.cpp
@@ -66,11 +66,11 @@ auto CommandQueue::submit(std::shared_ptr<Render::Fence> const& fence_) -> void
 	submitInfo.commandBufferCount = (uint32_t) commandBuffers.size();
 	submitInfo.pCommandBuffers = commandBuffers.data();
 
-	submitInfo.waitSemaphoreCount = cbBeginSemaphores.size();
+	submitInfo.waitSemaphoreCount = (uint32_t) cbBeginSemaphores.size();
 	submitInfo.pWaitSemaphores = cbBeginSemaphores.data();
 	submitInfo.pWaitDstStageMask = cbBeginPipelineFlags.data();
 
-	submitInfo.signalSemaphoreCount = cbEndSemaphores.size();
+	submitInfo.signalSemaphoreCount = (uint32_t) cbEndSemaphores.size();
 	submitInfo.pSignalSemaphores = cbEndSemaphores.data();
 
 	VkFence vulkanFence = VK_NULL_HANDLE;
